refactor(ST-43): modified file names to follow conventions and removed console statements
