--- modules/classroom/components/addStudentOnClassroom/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AddStudentOnClassroom";
