--- modules/classroom/components/addStudentOnClassroomForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AddStudentOnClassForm";
