+++ modules/classroom/components/addStudentToClassroom/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddStudentToClassroom";
