+++ modules/classroom/components/addStudentToClassroomForm/AddStudentToClassroomForm.types.tsx
@@ -1,0 +1,4 @@
+export interface AddStudentFormOptionProps {
+  id: string;
+  label: string;
+}
